test(streamfmt): separate stream regressions
The refactor-specific decoder and TUI polling cases made three broad test files
harder to navigate. Move those cases unchanged into focused files so the
provider protocol and view-lifecycle coverage remain easy to review.
